Autodetect fdt to use only at the first boot.
By default now fdt_file is "autodetect". Only in this case the env variable is automatically
set to the right dtb for the board in use. Calling saveenv() assures we do this only during
the first boot, and allows the users to set the fdt_file variable to any value avoiding any
automagic.